Removed HomeViewModel
*Added ProductsViewModel tests
