+++ CrashPasswordSystem/CrashPasswordSystem.UI/CrashPasswordSystem.UI.csproj
@@ -185,7 +185,6 @@
     <Compile Include="Modules\AddProducts\AddProductViewModel.cs" />
     <Compile Include="ViewModels\DetailViewModelBase.cs" />
     <Compile Include="ViewModels\Home2ViewModel.cs" />
-    <Compile Include="ViewModels\HomeViewModel.cs" />
     <Compile Include="ViewModels\LoginViewModel.cs" />
     <Compile Include="ViewModels\MainViewModel.cs" />
     <Compile Include="Modules\ProductDetails\ProductDetailsViewModel.cs" />
